+++ src/main/java/com/reckue/post/component/ServiceChecker.java
@@ -1,4 +1,4 @@
-package com.reckue.post.processor;
+package com.reckue.post.component;
 
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 @Component
 public class ServiceChecker {
 
-    protected static boolean isService(Class<?> beanClass) {
+    public boolean isService(Class<?> beanClass) {
         if (beanClass.isAnnotationPresent(Service.class)) {
             return true;
         } else {
@@ -14,7 +14,7 @@ public class ServiceChecker {
         }
     }
 
-    protected static boolean checkServiceAmongParents(Class<?>[] parents) {
+    public boolean checkServiceAmongParents(Class<?>[] parents) {
         boolean accumulator = false;
         for (Class<?> parent : parents) {
             accumulator = accumulator || isService(parent);
