+++ src/main/java/com/reckue/post/service/impl/CommentServiceImpl.java
@@ -8,7 +8,7 @@ import com.reckue.post.exception.model.post.PostNotFoundException;
 import com.reckue.post.model.Comment;
 import com.reckue.post.model.Node;
 import com.reckue.post.model.type.ParentType;
-import com.reckue.post.processor.annotation.NotNullArgs;
+import com.reckue.post.processor.notnull.NotNullArgs;
 import com.reckue.post.repository.CommentRepository;
 import com.reckue.post.repository.NodeRepository;
 import com.reckue.post.repository.PostRepository;
