+++ src/main/java/com/reckue/post/service/impl/RatingServiceImpl.java
@@ -7,7 +7,7 @@ import com.reckue.post.exception.model.rating.RatingNotFoundException;
 import com.reckue.post.exception.model.user.UserNotFoundException;
 import com.reckue.post.model.Post;
 import com.reckue.post.model.Rating;
-import com.reckue.post.processor.annotation.NotNullArgs;
+import com.reckue.post.processor.notnull.NotNullArgs;
 import com.reckue.post.repository.PostRepository;
 import com.reckue.post.repository.RatingRepository;
 import com.reckue.post.service.RatingService;
